fix: tabs and add tab content
