+++ src/index.js
@@ -1,7 +1,8 @@
 import Button from './mui/primitives/Button/Button';
 import ThemeProvider from './themes/ThemeProvider';
 import InputField from './mui/primitives/InputField/InputField';
+import Tabs from './mui/primitives/Tabs/Tabs';
 import { defaultTheme } from './themes/default/theme';
 import { kragleTheme } from './themes/kragle/theme';
 
-export { Button, ThemeProvider, InputField, defaultTheme, kragleTheme };
+export { Button, ThemeProvider, InputField, Tabs, defaultTheme, kragleTheme };
