+++ src/mui/primitives/Tabs/Tabs.module.scss
